+++ src/doc/content/changes.rst
@@ -1,4 +1,4 @@
-.. Copyright (C) 2005 - 2009  Eric Van Dewoestine
+.. Copyright (C) 2005 - 2010  Eric Van Dewoestine
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -15,6 +15,17 @@
 
 History of Changes
 ==================
+
+.. _1.5.5:
+
+1.5.5 (unreleased)
+----------------------
+
+Bug Fixes:
+  - Various bug fixes.
+
+Install:
+  - Added support for :ref:`automated installs <install-automated>`.
 
 .. _1.5.4: